+++ src/main/java/cleancode/studycafe/tobe/model/pass/StudyCafePassType.java
@@ -1,4 +1,4 @@
-package cleancode.studycafe.tobe.model;
+package cleancode.studycafe.tobe.model.pass;
 
 import java.util.Set;
