+++ src/main/java/cleancode/studycafe/tobe/model/pass/StudyCafeSeatPass.java
@@ -1,4 +1,6 @@
-package cleancode.studycafe.tobe.model;
+package cleancode.studycafe.tobe.model.pass;
+
+import cleancode.studycafe.tobe.model.pass.locker.StudyCafeLockerPass;
 
 public class StudyCafeSeatPass implements StudyCafePass {
