+++ src/main/java/cleancode/studycafe/tobe/model/pass/StudyCafeSeatPasses.java
@@ -1,4 +1,4 @@
-package cleancode.studycafe.tobe.model;
+package cleancode.studycafe.tobe.model.pass;
 
 
 import java.util.List;
