+++ src/main/java/cleancode/studycafe/tobe/model/pass/locker/StudyCafeLockerPasses.java
@@ -1,5 +1,7 @@
-package cleancode.studycafe.tobe.model;
+package cleancode.studycafe.tobe.model.pass.locker;
 
+
+import cleancode.studycafe.tobe.model.pass.StudyCafeSeatPass;
 
 import java.util.List;
 import java.util.Optional;
